# api-gateway/docs/swagger.yaml
...
openapi: 3.0.3
info:
  title: Granula API
  version: 1.0.0
  description: |
    # Granula API

    API для интеллектуального сервиса планирования ремонта и перепланировки квартир.

    ## Возможности

    - 🏠 **Загрузка планировок** - Загружайте изображения планировок для автоматического распознавания
    - 🤖 **AI распознавание** - Автоматическое распознавание комнат, стен и элементов
    - 🎨 **3D редактор** - Редактирование планировки в интерактивном 3D редакторе
    - ✅ **Проверка норм** - Автоматическая проверка на соответствие СНиП и ЖК РФ
    - 💡 **AI генерация** - Генерация вариантов планировки с помощью AI
    - 📋 **Заявки экспертам** - Подача заявок на консультации специалистов БТИ

    ## Аутентификация

    API использует JWT (JSON Web Tokens) для аутентификации.

    ### Получение токенов

    1. Зарегистрируйтесь через `POST /api/v1/auth/register`
    2. Или войдите через `POST /api/v1/auth/login`
    3. Получите `access_token` и `refresh_token`

    ### Использование токенов

    Передавайте Access Token в заголовке Authorization:
    ```
    Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    ```

    ### Обновление токенов

    Access Token действует 15 минут. Для обновления используйте `POST /api/v1/auth/refresh`

    ## Rate Limiting

    | Тип запроса | Лимит |
    |-------------|-------|
    | Анонимные | 60 req/min |
    | Аутентифицированные | 300 req/min |
    | AI endpoints | 30 req/min |

    При превышении лимита возвращается HTTP 429.

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен аутентификации. Получите через `/auth/login` или `/auth/register`.

        Формат: `Bearer <token>`

  # ===========================================================================
  # Schemas
  # ===========================================================================
  schemas:
...
    FloorPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        workspace_id:
          type: string
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440001
        name:
          type: string
          example: План квартиры - этаж 2
        description:
          type: string
          example: План второго этажа
        address:
          type: string
          example: г. Москва, ул. Тверская, д. 15
        status:
          type: string
          enum: [uploaded, processing, recognized, failed]
          description: |
            Статус обработки:
            - **uploaded** - Загружен, ожидает обработки
            - **processing** - Идет распознавание
            - **recognized** - Успешно распознан
            - **failed** - Ошибка распознавания
          example: recognized
        image_url:
          type: string
          format: uri
          example: https://storage.granula.ru/floorplans/fp-123.jpg
        thumbnail_url:
          type: string
          format: uri
          example: https://storage.granula.ru/thumbnails/fp-123-thumb.jpg
        recognition_result:
          $ref: '#/components/schemas/RecognitionResult'
        scene_id:
          type: string
          format: uuid
          description: ID связанной 3D сцены
          example: 770e8400-e29b-41d4-a716-446655440002
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RecognitionResult:
      type: object
      properties:
        rooms_count:
          type: integer
          description: Количество распознанных комнат
          example: 4
        total_area:
          type: number
          format: float
          description: Общая площадь (кв.м)
          example: 78.5
        living_area:
          type: number
          format: float
          description: Жилая площадь (кв.м)
          example: 52.3
        rooms:
          type: array
          items:
            $ref: '#/components/schemas/RecognizedRoom'
        confidence:
          type: number
          format: float
          description: Уверенность распознавания (0-1)
          example: 0.94
        processing_time_ms:
          type: integer
          description: Время обработки (мс)
          example: 1250
    RecognizedRoom:
      type: object
      properties:
        type:
          type: string
          enum: [living_room, bedroom, kitchen, bathroom, toilet, hallway, balcony, storage, other]
          example: bedroom
        name:
          type: string
          example: Спальня 1
        area:
          type: number
          format: float
          example: 18.5
        polygon:
          type: array
          items:
            $ref: '#/components/schemas/Point2D'
    Point2D:
      type: object
      properties:
        x:
          type: number
          format: float
          example: 120.5
        y:
          type: number
          format: float
          example: 80.3

    # -------------------------------------------------------------------------
    # Scene Schemas
    # -------------------------------------------------------------------------
...
paths:
...
  /floor-plans:
    post:
      tags: [floor-plans]
      summary: Загрузить планировку
      operationId: uploadFloorPlan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - workspace_id
                - name
                - file
              properties:
                workspace_id:
                  type: string
                  format: uuid
                name:
                  type: string
                description:
                  type: string
                address:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Планировка загружена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FloorPlan'
  /floor-plans/{id}/process:
    post:
      tags: [floor-plans]
      summary: Запустить распознавание
      operationId: processFloorPlan
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                create_scene:
                  type: boolean
                  default: true
                run_compliance:
                  type: boolean
                  default: true
      responses:
        '202':
          description: Обработка запущена
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: processing

  # ---------------------------------------------------------------------------
  # Compliance Endpoints
  # ---------------------------------------------------------------------------
...